requirements.txt
ultralytics 
opencv-python 
pillow
torch 
torchvision
torchaudio
timm